Moved CompositeRiskModel [skip hazardlib][demos]
