+++ openquake/commonlib/tests/readinput_test.py
@@ -22,10 +22,10 @@ import unittest.mock as mock
 import unittest
 from io import BytesIO
 
-from openquake.baselib import general, performance, datastore
+from openquake.baselib import general, datastore
 from openquake.hazardlib import InvalidFile
 from openquake.risklib import asset
-from openquake.risklib.riskinput import ValidationError
+from openquake.risklib.riskmodels import ValidationError
 from openquake.commonlib import readinput
 from openquake.qa_tests_data.classical import case_1, case_2, case_21
 from openquake.qa_tests_data.event_based import case_16
